requirements.txt
azure-devops==6.0.0b1
msrest>=0.6.0,<0.7.0
requests>=2.28.0
setuptools
